Time: 0 ms (100.00%) | Memory: 8.6 MB (44.56%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double myPow(double x, int n) {
-        long binary = n;
+        long long binary = n;
         double ans = 1;
         if (binary<0){
             x = 1/x;
